split views.py into seperate files inside a views folder
